+++ sqlitemulti/HISTORY.rst
@@ -6,3 +6,13 @@ History
 ------------------
 
 * First release on PyPI.
+
+0.0.2 (2019-06-05)
+------------------
+
+* First usable release.
+* Processes or Threads
+* Pooled tasks
+* Benchmarks
+* Explicit free of ending threads
+* Experimental garbage collector for results queues